feat: docker backend implementation
