+++ .dockerignore
@@ -1,0 +1,12 @@
+node_modules
+npm-debug.log
+Dockerfile*
+.dockerignore
+.git
+.gitignore
+contacts.db
+*.sqlite
+*.sqlite-journal
+coverage
+logs
+.env
