+++ .env.example
@@ -1,0 +1,8 @@
+# Puerto de la aplicación
+PORT=3000
+
+# Ruta del archivo de base de datos dentro del contenedor (persistente vía volumen)
+DB_FILE=/data/contacts.db
+
+# NODE_ENV puede ser development o production
+NODE_ENV=development
