+++ backend/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "u-la-salle-contact-form-backend",
+  "version": "1.0.0",
+  "description": "Backend API para formulario de contacto U La Salle",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js",
+    "dev": "nodemon server.js",
+    "start:docker": "node server.js",
+  "test": "echo 'No tests yet' && exit 0",
+  "db:query": "node query.js",
+  "db:stats": "node query.js stats",
+  "db:last": "node query.js last 10"
+  },
+  "dependencies": {
+    "express": "^4.18.2",
+    "sqlite3": "^5.1.6",
+    "cors": "^2.8.5",
+    "body-parser": "^1.20.2",
+    "dotenv": "^16.3.1"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.1"
+  },
+  "license": "MIT"
+}
