+++ package.json
@@ -1,23 +1,9 @@
 {
-  "name": "u-la-salle-contact-form",
-  "version": "1.0.0",
-  "description": "Simple web application with contact form for U La Salle",
-  "main": "server.js",
+  "name": "u-la-salle-contact-form-root",
+  "private": true,
+  "description": "Archivo placeholder. El backend se movió a ./backend",
+  "workspaces": ["backend"],
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": ["contact-form", "web-application", "node.js", "sqlite"],
-  "author": "U La Salle",
-  "license": "MIT",
-  "dependencies": {
-    "express": "^4.18.2",
-    "sqlite3": "^5.1.6",
-    "cors": "^2.8.5",
-    "body-parser": "^1.20.2"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.1"
+    "start": "echo 'Backend se movió a ./backend. Ejecuta: cd backend && npm start' && exit 1"
   }
 }
